Changes for Windows + text bugfix
There was an issue of text not being displayed on some machines. The bug
turned out to be a simple issue of one of the text shader's uniforms not
being set (the world shader's uniform was being set instead). Presumably
since the same vert shader was being used for both programs, some drivers
shared the uniforms, which hid the bug on those machines.