UI: Screen::Inform special case removed
All UIElements are now assumed by Screen::Inform to be a part of
the render list (UIViewports were previously assumed to not be a
part of this list). This saves a lot of headache in terms of users
no longer needing to know when a UIElement is a part of the render
list.

Future changes for "priority" of UIElements can contain a special
case priority (-1 or 0), which the DrawSystem can ignore.