Cleaned up Random code (proper use of <random> distributions)
The previous Random implementation used modulus in order to generate a
range smaller than that of the random engine. There are a number of
well-documented issues with doing this, and therefore this change uses
distributions instead.

The code currently constructs a distribution instance every time
any Random::GetRand is called. While distributions do not take nearly
as much time to construct as random engines, future changes may
involve the storage of common/recently-used distribution objects in
order to avoid this overhead.