build: Added basic framework for build_tool
In order to supplement C++ with reflection functionality, tetrad-engine
will employ the use of a tool to be run prior to project compilation,
which will involve parsing out various pieces of information from C++
header files and writing that information out as C++ files themselves.

This change provides the basic framework for the tool. It currently
does not actually parse the C++ headers, but simply keeps track of which
files should be considered.