ecs: Use consistent include convention
Includes should be ordered as follows:
  1. Matching header (for cpp files)
  2. Standard library and system headers
  3. Third-party headers
  4. Tetrad headers

Each section that exists should have a space between it and an
adjacent section. Only section 2 should use angle brackets as
opposed to quotes. Sections 1 and 4 should refer to the file
relative to the tetrad src/ directory.

This change applies such rules to the ecs package.