Add component type annotation for build_tool
build_tool needs information about components in order to generate
reflection code, as well as explicit template instantiations. This
change adds a COMPONENT macro used to indicate a component
definition. Currently the macro has no options, but will still be
useful for build_tool to keep track of which components a project
defines.