build: Add functionality to compile protobufs
This change adds functionality to the build system to compile
protobufs prior to building other targets. Any file added to the
PROTOS list will be compiled, with the output files placed in the
generated directory.

The protobuf git repo is added to the external directory, such that
client systems do not have to have protoc installed in order for
builds to succeed. While this does slow down the first build,
subsequent builds should never have to rebuild any of the protobuf
targets.