build: Run build_tool during cmake configuration
5d9e82c moves the running of build_tool to build time. However, cmake
requires that referenced generated files exist prior to build time,
meaning that the new functionality breaks cmake generation for new
projects. This change causes build_tool to be run both that generation
time and at build time. While this means that build_tool will be run
twice on builds that cause cmake to be triggered at build time,
build_tool is designed to exit quickly when it has no work to do.

TEST=Successfully generated new cmake build.